New commit

+++ database/migrations/2026_02_12_102647_add_foreign_key_to_rented_table_vendor_id.php
@@ -12,10 +12,12 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('rented', function (Blueprint $table) {
-             $table->foreign('vendor_id')
-                  ->references('id')
-                  ->on('vendor_details')
-                  ->onDelete('cascade');
+            if (Schema::hasColumn('rented', 'vendor_id')) {
+                $table->foreign('vendor_id')
+                      ->references('id')
+                      ->on('vendor_details')
+                      ->onDelete('cascade');
+            }
         });
     }
